+++ packages/docs/components/Menu/stories/mega-menu/app.js
@@ -1,11 +1,12 @@
 import { Base, createApp } from '@studiometa/js-toolkit';
-import { Menu } from '@studiometa/ui';
+import { Menu, Action } from '@studiometa/ui';
 
 class App extends Base {
   static config = {
     name: 'App',
     components: {
       Menu,
+      Action,
     },
   };
 }